requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
discord.py==2.4.0
faster-whisper==1.1.0
